Extracts screen code into its own module
This continues the movement for smaller, contained modules.

I am trying to set up the future refactoring that we need to do to be
much simpler.
